Register/requirements.txt
React
Tailwind
flask
bcrypt
pyjwt

